fix: consolidate Supabase middleware to single source of truth
CRITICAL FIX: Home page redirecting all users to /login

Root Cause:
- Duplicate Supabase clients in lib/supabase/ AND utils/supabase/
- middleware.ts imported WRONG middleware (missing auth logic)
- utils/supabase/middleware.ts had production code but wasn't used

Solution:
- Moved production middleware to lib/supabase/middleware.ts
- Deleted duplicate utils/supabase/ directory
- Single source of truth: lib/supabase/ (21 files verified)

Production Middleware Features (Now Active):
- Public routes accessible (/, /features, /pricing, /about, /blog, /faq, /contact)
- Protected routes enforce auth (/manage/*)
- Payment gating with Stripe subscription validation
- Role-based redirects (TENANT → /tenant, OWNER → /manage)
- Proper auth with getUser() validation

Fixes:
- ✅ Home page accessible to all users
- ✅ Marketing pages accessible
- ✅ Auth routes redirect authenticated users
- ✅ Protected routes require authentication
- ✅ Payment-required routes enforce subscription

Verified with Serena MCP:
- 21 files use @/lib/supabase ✅
- 0 files use @/utils/supabase ✅
