+++ apps/backend/package.json
@@ -45,6 +45,8 @@
 		"@types/passport-jwt": "4.0.1",
 		"@types/pg": "8.15.5",
 		"cache-manager": "7.2.4",
+		"class-transformer": "0.5.1",
+		"class-validator": "0.14.2",
 		"compression": "^1.8.1",
 		"cookie-parser": "^1.4.7",
 		"exceljs": "4.4.0",
