--- apps/frontend/src/utils/supabase/client.ts
@@ -1,9 +1,0 @@
-import type { Database } from '@repo/shared/types/supabase-generated'
-import { createBrowserClient } from '@supabase/ssr'
-
-export function createClient() {
-	return createBrowserClient<Database>(
-		process.env.NEXT_PUBLIC_SUPABASE_URL!,
-		process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-	)
-}
